+++ CHANGELOG.md
@@ -6,9 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.4.0-beta.1] - 2022-11-26
+
 ## Added
 
 - **! Added `Table.get_row_where(self: Self@Table, attribute_name: str, attribute_value: Any) -> tuple[Any, ...]`.**
+- **! Added `Table.get_rows_where(self: Self@Table, attribute_name: str, attribute_value: Any, allow_empty: bool = False) -> list[tuple[Any, ...]]`.**
 
 ## [0.3.0] - 2022-11-25